+++ src/misc.ts
@@ -1,1 +1,2 @@
 export type NewLineType = '\r\n'|'\n';
+export type SplitSourceText = [string, string|null, string|null];